we don't need newTrades collections in portfolio anymore & now we use Long instead of Integer for quantity

+++ src/main/java/com/highbridge/trade_allocation/domain/Account.java
@@ -37,15 +37,15 @@ public class Account {
         return stockTargetPercent.get(stock).of(capital);
     }
 
-    Integer targetMarketQuantity(String stock, Money price) {
-        return targetMarketValue(stock).divide(price).getAmount().intValue();
+    Long targetMarketQuantity(String stock, Money price) {
+        return targetMarketValue(stock).divide(price).getAmount().longValue();
     }
 
-    Integer currentQuantity(String stock) {
+    Long currentQuantity(String stock) {
         return holdings.get(stock).quantity();
     }
 
-    Integer quantityCanBeAdded(String stock, Money price) {
+    Long quantityCanBeAdded(String stock, Money price) {
         return targetMarketQuantity(stock, price) - currentQuantity(stock);
     }
