+++ src/main/java/com/highbridge/trade_allocation/domain/Trade.java
@@ -6,22 +6,22 @@ import static com.highbridge.trade_allocation.domain.Trade.BuyOrSell.*;
 
 public class Trade {
     private final String stock;
-    private final Integer quantity;
+    private final Long quantity;
     private final Money price;
     private final BuyOrSell buyOrSell;
 
-    private Trade(String stock, Integer quantity, Money price, BuyOrSell buyOrSell) {
+    private Trade(String stock, Long quantity, Money price, BuyOrSell buyOrSell) {
         this.stock = stock;
         this.quantity = quantity;
         this.price = price;
         this.buyOrSell = buyOrSell;
     }
 
-    static Trade buy(String stock, Integer quantity, Money price) {
+    static Trade buy(String stock, Long quantity, Money price) {
         return new Trade(stock, quantity, price, Buy);
     }
 
-    static Trade sell(String stock, Integer quantity, Money price) {
+    static Trade sell(String stock, Long quantity, Money price) {
         return new Trade(stock, quantity, price, Sell);
     }
 
@@ -29,7 +29,7 @@ public class Trade {
         return stock;
     }
 
-    Integer singedQuantity() {
+    Long singedQuantity() {
         return this.buyOrSell == Buy ? quantity : -quantity;
     }
